refactor(chat): move /task dispatch into chat_execution_service — #1483
Decompose the CC-82 execute_parallel_task monster (AC #2). The router keeps only
the request-boundary guards (container 404/503, #1672 resume IDOR gate, #1068
timeout normalization — the redis-via-router helper stays here per RD10) and a
thin call that maps ChatDispatchError → HTTPException / ChatAdmissionReplay →
409/200. Everything else — derive/spoof (403), idempotency begin/replay, file
upload (502), row+activity creation, and the async/sync fork — moves to
chat_execution_service, each function ≤ CC 20 / ≤ 150 lines:

  dispatch_parallel_task, derive_source_and_trigger, process_task_file_uploads,
  create_task_execution_and_activities, _acquire_task_capacity, _map_task_failure,
  _dispatch_async, _dispatch_sync{,_immediate,_backlog}, run_async_task,
  complete_collaboration_activity, finalize_self_task (kept WHOLE, RD9).

execute_parallel_task: CC 82 → 13; routers/chat.py: 2756 → 1132 lines. The sync
paths still delegate to task_execution_service.execute_task (never a 2nd applier,
RD1). Circuit/ephemeral 503/410 + FAILED-write mirrored in the service (single
place per path — no double-write, RD-E12). #1578 reserved-event tag flows through
every sink; #914 queued-202 shape, #1444 chat_persist_failed, and the upload-502
no-idempotency-fail quirk (RD11) preserved. OpenAPI byte-identical.

- Repoint the production importer backlog_service.py:265 →
  services.chat_execution_service.run_async_task (kept LAZY — breaks a real cycle).
- The chat router no longer holds a WebSocket manager (all broadcasts moved to the
  two services, each with its own setter — §7 minimize-WS-globals); main.py wiring
  updated. Unused router imports trimmed.
- Tests repointed in-commit to the new collaborator modules (module-identity):
  the /task char suite, test_946, test_1578, test_backlog (fake-module + AST guard),
  test_1332, test_1457, test_async_task_persistence, test_1444, test_1672.
