+++ src/backend/main.py
@@ -54,7 +54,7 @@ from routers.credentials import router as credentials_router
 from routers.templates import router as templates_router
 from routers.sharing import router as sharing_router, set_websocket_manager as set_sharing_ws_manager
 from routers.mcp_keys import router as mcp_keys_router
-from routers.chat import router as chat_router, set_websocket_manager as set_chat_ws_manager
+from routers.chat import router as chat_router
 from services.chat_persistence_service import set_websocket_manager as set_chat_persistence_ws_manager
 from services.chat_execution_service import set_websocket_manager as set_chat_execution_ws_manager
 from routers.sessions import router as sessions_router  # SESSION_TAB_2026-04 Phase 2
@@ -244,9 +244,8 @@ set_agents_filtered_ws_manager(filtered_manager)
 set_agent_rename_ws_manager(manager)
 set_agent_rename_filtered_ws_manager(filtered_manager)
 set_sharing_ws_manager(manager)
-set_chat_ws_manager(manager)
 set_chat_persistence_ws_manager(manager)  # #1483: chat_response_ready broadcast
-set_chat_execution_ws_manager(manager)    # #1483: agent_collaboration broadcast
+set_chat_execution_ws_manager(manager)    # #1483: agent_collaboration + self_task broadcasts
 set_public_links_ws_manager(manager)
 set_notifications_ws_manager(manager)
 set_notifications_filtered_ws_manager(filtered_manager)
