+++ src/backend/services/chat_execution_service.py
@@ -24,10 +24,12 @@ the exact (status_code, detail, headers) the thin router maps 1:1. The lone
 — a defensive *re-raise* of an exception ``subscription_auto_switch`` itself
 raises (propagate-unchanged), never HTTP construction.
 """
+import asyncio
 import httpx
 import json
 import logging
 import uuid
+from collections import namedtuple
 from datetime import datetime
 from typing import Optional
 
@@ -36,6 +38,7 @@ from fastapi import HTTPException  # only for the defensive SUB-003 re-raise
 from models import (
     User,
     ChatMessageRequest,
+    ParallelTaskRequest,
     ActivityType,
     ActivityState,
     TaskExecutionStatus,
@@ -48,15 +51,40 @@ from services.model_context import DEFAULT_CONTEXT_WINDOW
 from services.task_execution_service import (
     _compute_context_used,
     agent_post_with_retry,
+    get_task_execution_service,
+    dispatch_breaker_active,
 )
+from services.capacity_manager import (
+    CapacityFull,
+    CircuitOpen,
+    EphemeralBudgetExhausted,
+    PersistentTaskPayload,
+    get_capacity_manager,
+)
+from services.upload_service import (
+    process_file_uploads,
+    decode_web_file,
+    WEB_MAX_FILES,
+    WEB_MAX_FILE_SIZE,
+    WEB_MAX_IMAGE_SIZE,
+    WEB_MAX_TOTAL_IMAGE_SIZE,
+)
+from services.sync_waiter import signal_sync_waiter, wait_for_sync_terminal
+from services.event_dispatch_service import (
+    RESERVED_EVENT_TRIGGER,
+    RESERVED_EVENT_TRIGGER_HEADER_VALUE,
+    verify_internal_dispatch_secret,
+)
+from services import idempotency_service
+from services import dispatch_admission_service
+from services import chat_persistence_service
 from services.platform_prompt_service import (
     ExecutionContext,
     compose_system_prompt,
     get_platform_system_prompt,
     is_execution_context_enabled,
 )
-from services import idempotency_service
-from services.chat_signals import ChatExecutionContext, ChatDispatchError
+from services.chat_signals import ChatExecutionContext, ChatAdmissionReplay, ChatDispatchError
 from utils.credential_sanitizer import sanitize_dict, sanitize_execution_log, sanitize_response
 from utils.helpers import utc_now_iso
 
@@ -639,3 +667,802 @@ async def run_chat_turn(
         # success path, where complete() already finalized it).
         if not idem_done:
             idempotency_service.fail(idem)
+
+
+# ===========================================================================
+# /task dispatch lifecycle (#1483). NOTE: this service is NEVER a /task terminal
+# applier — the sync-immediate and backlog paths delegate to
+# task_execution_service.execute_task (the single applier). This section owns
+# only the endpoint normalization + dispatch orchestration + chat-specific
+# post-processing (RD1).
+# ===========================================================================
+
+_TaskDerivation = namedtuple(
+    "_TaskDerivation", ["triggered_by", "is_self_task", "reserved_event_dispatch"]
+)
+
+
+async def run_async_task(
+    agent_name: str,
+    request: ParallelTaskRequest,
+    execution_id: str,
+    collaboration_activity_id: Optional[str],
+    x_source_agent: Optional[str],
+    user_id: Optional[int] = None,
+    user_email: Optional[str] = None,
+    subscription_id: Optional[str] = None,
+    is_self_task: bool = False,
+    self_task_activity_id: Optional[str] = None,
+    images: Optional[list] = None,
+    dispatch_gate_checked: bool = False,
+    triggered_by_override: Optional[str] = None,
+):
+    """
+    Async /task background wrapper (issue #95).
+
+    Delegates the full execution lifecycle to TaskExecutionService (single path
+    for slot / activity / sanitization / retry / release) and layers on the
+    chat-endpoint-specific post-task side effects:
+      - authenticated chat_session persistence (THINK-001)
+      - chat_response_ready WebSocket broadcast
+      - collaboration activity completion (agent-to-agent call)
+
+    Caller (the async branch) has already pre-acquired the capacity slot so that
+    429-at-capacity is returned synchronously. The service releases the slot in
+    its finally block.
+    """
+    start_time = datetime.utcnow()
+    task_service = get_task_execution_service()
+    # #1578: a reserved agent.task.* dispatch persists triggered_by="event" (the
+    # recursion-break sentinel); every other async task keeps the derived value.
+    triggered_by = triggered_by_override or ("agent" if x_source_agent else "manual")
+
+    # Outer try/finally so a sync long-poll waiter (issue #498) is always
+    # signaled even if the post-task side effects below raise.
+    result = None
+    chat_session_id = None
+    try:
+        result = await task_service.execute_task(
+            agent_name=agent_name,
+            message=request.message,
+            triggered_by=triggered_by,
+            source_user_id=user_id,
+            source_user_email=user_email,
+            source_agent_name=x_source_agent,
+            model=request.model,
+            timeout_seconds=request.timeout_seconds,
+            resume_session_id=request.resume_session_id,
+            allowed_tools=request.allowed_tools,
+            system_prompt=request.system_prompt,
+            execution_id=execution_id,
+            subscription_id=subscription_id,
+            parent_activity_id=collaboration_activity_id,
+            extra_activity_details={
+                "parallel_mode": True,
+                "async_mode": True,
+                "model": request.model,
+                "timeout_seconds": request.timeout_seconds,
+            },
+            slot_already_held=True,  # Router pre-acquired to preserve 429-upfront contract
+            images=images or [],
+            dispatch_gate_checked=dispatch_gate_checked,  # #526: True when router gated at acquire()
+        )
+
+        execution_time_ms = int((datetime.utcnow() - start_time).total_seconds() * 1000)
+
+        # Post-task side effects (each guarded + self-isolating; see helpers).
+        chat_session_id = await chat_persistence_service.persist_and_broadcast_chat_session(
+            agent_name=agent_name, request=request, result=result,
+            execution_id=execution_id, user_id=user_id, user_email=user_email,
+            subscription_id=subscription_id, execution_time_ms=execution_time_ms,
+        )
+        await complete_collaboration_activity(
+            collaboration_activity_id, result, execution_id, execution_time_ms,
+        )
+        await finalize_self_task(
+            is_self_task=is_self_task, self_task_activity_id=self_task_activity_id,
+            agent_name=agent_name, request=request, result=result,
+            execution_id=execution_id, user_id=user_id, user_email=user_email,
+            execution_time_ms=execution_time_ms,
+        )
+
+        logger.info(
+            f"[Task Async] Completed background task for agent '{agent_name}', "
+            f"execution_id={execution_id}, status={result.status}"
+        )
+    finally:
+        # Issue #498: signal any sync HTTP caller waiting on this execution.
+        # No-op when no waiter is registered (the common async path).
+        signal_sync_waiter(execution_id, result, chat_session_id)
+
+
+async def complete_collaboration_activity(
+    collaboration_activity_id, result, execution_id, execution_time_ms,
+):
+    """Post-task block 2: complete the agent-to-agent collaboration activity.
+    No-op when there is no collaboration activity; self-isolating on error."""
+    if not collaboration_activity_id:
+        return
+    try:
+        await activity_service.complete_activity(
+            activity_id=collaboration_activity_id,
+            # #1332: a cancelled collaboration turn reads as CANCELLED, not FAILED.
+            status=activity_state_for_terminal(result.status),
+            details={
+                "response_length": len(result.response or ""),
+                "execution_time_ms": execution_time_ms,
+                "execution_id": execution_id,
+            },
+            error=(result.error if result.status == TaskExecutionStatus.FAILED else None),
+        )
+    except Exception as e:
+        logger.warning(f"[Task Async] collaboration activity completion failed: {e}")
+
+
+async def finalize_self_task(
+    *, is_self_task, self_task_activity_id, agent_name, request, result,
+    execution_id, user_id, user_email, execution_time_ms,
+):
+    """Post-task block 3 (SELF-EXEC-001): complete the self-task activity,
+    inject the result into the originating chat session when requested, and
+    broadcast the completion event. No-op unless this is a self-task. Kept WHOLE
+    (RD9): its shared activity_status/result must not fracture across modules."""
+    if not (is_self_task and self_task_activity_id):
+        return
+
+    # #1332: a cancelled self-task turn reads as CANCELLED, not FAILED.
+    activity_status = activity_state_for_terminal(result.status)
+
+    # Complete the self-task activity
+    try:
+        await activity_service.complete_activity(
+            activity_id=self_task_activity_id,
+            status=activity_status,
+            details={
+                "response_length": len(result.response or ""),
+                "execution_time_ms": execution_time_ms,
+                "execution_id": execution_id,
+                "inject_result": request.inject_result,
+            },
+            error=(result.error if result.status == TaskExecutionStatus.FAILED else None),
+        )
+    except Exception as e:
+        logger.warning(f"[Task Async] self-task activity completion failed: {e}")
+
+    # Inject result into chat session if requested
+    if request.inject_result and request.chat_session_id and result.status == TaskExecutionStatus.SUCCESS:
+        try:
+            # Validate session exists and belongs to user
+            session = db.get_chat_session(request.chat_session_id)
+            if session and session.user_id == user_id:
+                # Add self-task result as a chat message
+                db.add_chat_message(
+                    session_id=request.chat_session_id,
+                    agent_name=agent_name,
+                    user_id=user_id,
+                    user_email=user_email or "",
+                    role="assistant",
+                    content=result.response or "",
+                    cost=result.cost,
+                    context_used=result.context_used,
+                    context_max=result.context_max,
+                    execution_time_ms=execution_time_ms,
+                    source="self_task",  # Mark as self-task result
+                )
+                logger.info(f"[Self-Task] Injected result into chat session {request.chat_session_id}")
+            else:
+                logger.warning(f"[Self-Task] Cannot inject result: session {request.chat_session_id} not found or not owned by user")
+        except Exception as e:
+            logger.warning(f"[Self-Task] Failed to inject result into chat session: {e}")
+
+    # Broadcast self-task completion event
+    if _websocket_manager:
+        try:
+            await _websocket_manager.broadcast(json.dumps({
+                "type": "agent_activity",
+                "agent_name": agent_name,
+                "activity_type": "self_task",
+                # #1332: mirror the activity DB state (cancelled stays cancelled)
+                # so the WS event and the persisted row never disagree.
+                "activity_state": activity_status.value,
+                "action": f"Background task completed",
+                "timestamp": utc_now_iso(),
+                "details": {
+                    "execution_id": execution_id,
+                    "chat_session_id": request.chat_session_id,
+                    "cost_usd": result.cost,
+                    "execution_time_ms": execution_time_ms,
+                    "response_preview": (result.response or "")[:200],
+                    "inject_result": request.inject_result,
+                    "result_injected": request.inject_result and request.chat_session_id is not None,
+                }
+            }))
+        except Exception as e:
+            logger.warning(f"[Self-Task] WebSocket broadcast failed: {e}")
+
+
+def derive_source_and_trigger(
+    *, name, x_source_agent, x_via_mcp, x_event_trigger, x_internal_secret, current_user
+) -> "_TaskDerivation":
+    """SELF-EXEC-001 spoof guard (403), self-task detection, triggered_by
+    derivation, and the #1578 reserved-event tag (internal-secret gated, C-003).
+    HTTP-free — the spoof guard raises ChatDispatchError(403)."""
+    # SELF-EXEC-001: verify X-Source-Agent matches the MCP key's agent scope.
+    if x_source_agent and current_user.agent_name:
+        if x_source_agent != current_user.agent_name:
+            raise ChatDispatchError(
+                403,
+                f"Source agent header '{x_source_agent}' doesn't match API key scope '{current_user.agent_name}'",
+            )
+
+    is_self_task = (x_source_agent is not None and x_source_agent == name)
+
+    if x_source_agent:
+        triggered_by = "self_task" if is_self_task else "agent"
+    elif x_via_mcp:
+        triggered_by = "mcp"
+    else:
+        triggered_by = "manual"
+
+    # #1578 recursion-break: honored ONLY with a valid backend-internal secret.
+    reserved_event_dispatch = (
+        x_event_trigger == RESERVED_EVENT_TRIGGER_HEADER_VALUE
+        and verify_internal_dispatch_secret(x_internal_secret)
+    )
+    if reserved_event_dispatch:
+        triggered_by = RESERVED_EVENT_TRIGGER
+
+    return _TaskDerivation(
+        triggered_by=triggered_by,
+        is_self_task=is_self_task,
+        reserved_event_dispatch=reserved_event_dispatch,
+    )
+
+
+async def process_task_file_uploads(*, request, name, container, current_user) -> list:
+    """(#364) File upload processing before the async/sync fork. Mutates
+    request.message with the file block; returns the decoded image data. Raises
+    ChatDispatchError(502) when ALL writes fail — and, matching the original,
+    does NOT touch the idempotency claim on that path (RD11; the caller has
+    already begun the claim, which is intentionally left in place)."""
+    image_data: list = []
+    if request.files:
+        uploader = current_user.email or current_user.username
+        raw_files = [
+            {
+                "name": f.name,
+                "mimetype": f.mimetype,
+                "size": f.size,
+                "data": decode_web_file(f.dict()),
+                "id": f"f{i}",
+            }
+            for i, f in enumerate(request.files)
+        ]
+        file_descs, _upload_dir, all_writes_failed, image_data = await process_file_uploads(
+            raw_files=raw_files,
+            agent_name=name,
+            container=container,
+            session_id=str(current_user.id),
+            uploader=uploader,
+            source="web",
+            max_files=WEB_MAX_FILES,
+            max_file_size=WEB_MAX_FILE_SIZE,
+            max_image_size=WEB_MAX_IMAGE_SIZE,
+            max_total_image_size=WEB_MAX_TOTAL_IMAGE_SIZE,
+        )
+        if all_writes_failed:
+            raise ChatDispatchError(
+                502, "File upload failed: could not write to agent workspace."
+            )
+        if file_descs:
+            file_block = "\n".join(file_descs)
+            request.message = f"{request.message}\n\n{file_block}"
+    return image_data
+
+
+async def create_task_execution_and_activities(
+    *, request, name, current_user, x_source_agent, x_mcp_key_id, x_mcp_key_name,
+    triggered_by, is_self_task, idem,
+):
+    """Create the execution record (#95/#96), attach the idempotency claim, and
+    track the collaboration / self-task activity (mirrors the /chat pattern).
+    Returns (execution_id, subscription_id, collaboration_activity_id,
+    self_task_activity_id)."""
+    # SUB-004: subscription snapshot at creation time (best-effort).
+    try:
+        subscription_id = db.get_agent_subscription_id(name)
+    except Exception:
+        subscription_id = None
+
+    execution = db.create_task_execution(
+        agent_name=name,
+        message=request.message,
+        triggered_by=triggered_by,
+        source_user_id=current_user.id,
+        source_user_email=current_user.email or current_user.username,
+        source_agent_name=x_source_agent,
+        source_mcp_key_id=x_mcp_key_id,
+        source_mcp_key_name=x_mcp_key_name,
+        model_used=request.model,
+        subscription_id=subscription_id,
+    )
+    execution_id = execution.id if execution else None
+    idempotency_service.attach_execution(idem, execution_id)
+
+    collaboration_activity_id = None
+    self_task_activity_id = None
+    if x_source_agent:
+        if is_self_task:
+            self_task_activity_id = await activity_service.track_activity(
+                agent_name=name,  # Activity belongs to the agent running the self-task
+                activity_type=ActivityType.SELF_TASK,
+                user_id=current_user.id,
+                triggered_by="self_task",
+                related_execution_id=execution_id,
+                details={
+                    "agent_name": name,
+                    "action": "self_task",
+                    "message_preview": request.message[:100],
+                    "execution_id": execution_id,
+                    "parallel_mode": True,
+                    "inject_result": request.inject_result,
+                    "chat_session_id": request.chat_session_id,
+                }
+            )
+            if _websocket_manager:
+                await _websocket_manager.broadcast(json.dumps({
+                    "type": "agent_activity",
+                    "agent_name": name,
+                    "activity_type": "self_task",
+                    "activity_state": "started",
+                    "action": f"Background task: {request.message[:50]}...",
+                    "timestamp": utc_now_iso(),
+                    "details": {
+                        "execution_id": execution_id,
+                        "chat_session_id": request.chat_session_id,
+                        "message_preview": request.message[:100],
+                        "inject_result": request.inject_result,
+                    }
+                }))
+        else:
+            await broadcast_collaboration_event(
+                source_agent=x_source_agent,
+                target_agent=name,
+                action="parallel_task"
+            )
+            collaboration_activity_id = await activity_service.track_activity(
+                agent_name=x_source_agent,  # Activity belongs to source agent (the caller)
+                activity_type=ActivityType.AGENT_COLLABORATION,
+                user_id=current_user.id,
+                triggered_by="agent",
+                related_execution_id=execution_id,
+                details={
+                    "source_agent": x_source_agent,
+                    "target_agent": name,
+                    "action": "parallel_task",
+                    "message_preview": request.message[:100],
+                    "execution_id": execution_id,
+                    "parallel_mode": True
+                }
+            )
+
+    return execution_id, subscription_id, collaboration_activity_id, self_task_activity_id
+
+
+def _circuit_open_dispatch_error(name, execution_id, exc) -> ChatDispatchError:
+    """Byte-identical mirror of routers.chat._raise_circuit_open_503, but returns
+    the domain error instead of raising HTTP — closes the pre-created row FAILED
+    (the /task path has a row) then builds the 503 + X-Circuit-Open/Retry-After.
+    The router's _raise_circuit_open_503 is used ONLY on /chat (no row), so the
+    FAILED-write is never double-written (RD-E12)."""
+    if execution_id:
+        try:
+            db.update_execution_status(
+                execution_id=execution_id,
+                status=TaskExecutionStatus.FAILED,
+                error="circuit_open: agent unhealthy (dispatch breaker open)",
+            )
+        except Exception as e:
+            logger.warning(
+                f"[Chat] Failed to mark execution {execution_id} FAILED on circuit open: {e}"
+            )
+    retry_after = max(0, int(exc.retry_after_seconds))
+    return ChatDispatchError(
+        503,
+        {"error": "circuit_open", "retry_after_seconds": retry_after},
+        headers={"X-Circuit-Open": "true", "Retry-After": str(retry_after)},
+    )
+
+
+def _ephemeral_dispatch_error(name, execution_id, exc) -> ChatDispatchError:
+    """Byte-identical mirror of routers.chat._raise_ephemeral_exhausted_410 (with
+    the /task row FAILED-write)."""
+    if execution_id:
+        try:
+            db.update_execution_status(
+                execution_id=execution_id,
+                status=TaskExecutionStatus.FAILED,
+                error=f"ephemeral_exhausted: ghost agent budget spent ({exc.reason})",
+            )
+        except Exception as e:
+            logger.warning(
+                f"[Chat] Failed to mark execution {execution_id} FAILED on ephemeral exhaustion: {e}"
+            )
+    return ChatDispatchError(
+        410,
+        {
+            "error": f"Ephemeral agent '{name}' budget is spent ({exc.reason}); it is being discarded.",
+            "code": "ephemeral_budget_exhausted",
+        },
+    )
+
+
+async def _acquire_task_capacity(
+    *, mode_label, name, request, execution_id, triggered_by, collaboration_activity_id,
+    is_self_task, self_task_activity_id, user_id, user_email, subscription_id,
+    x_source_agent, x_mcp_key_id, x_mcp_key_name, idem,
+):
+    """Pre-acquire the /task capacity slot (queue_persistent overflow) shared by
+    the async and sync branches. On a deny it releases the idempotency claim and
+    raises a ChatDispatchError (429 capacity-full / 503 circuit / 410 ephemeral).
+    Returns (cap_result, effective_timeout)."""
+    capacity = get_capacity_manager()
+    max_parallel_tasks = db.get_max_parallel_tasks(name)
+    cb_enabled = dispatch_breaker_active(name)  # #526: combined global + per-agent gate
+    effective_timeout = request.timeout_seconds
+    if effective_timeout is None:
+        effective_timeout = db.get_execution_timeout(name)
+
+    try:
+        cap_result = await capacity.acquire(
+            agent_name=name,
+            execution_id=execution_id or f"temp-{datetime.utcnow().timestamp()}",
+            max_concurrent=max_parallel_tasks,
+            message_preview=request.message[:100] if request.message else "",
+            timeout_seconds=effective_timeout,
+            overflow_policy="queue_persistent",
+            breaker_enabled=cb_enabled,
+            overflow_payload=PersistentTaskPayload(
+                request=request,
+                effective_timeout=effective_timeout,
+                user_id=user_id,
+                user_email=user_email,
+                subscription_id=subscription_id,
+                x_source_agent=x_source_agent,
+                x_mcp_key_id=x_mcp_key_id,
+                x_mcp_key_name=x_mcp_key_name,
+                triggered_by=triggered_by,
+                collaboration_activity_id=collaboration_activity_id,
+                is_self_task=is_self_task,
+                self_task_activity_id=self_task_activity_id,
+            ),
+        )
+    except CapacityFull:
+        # Both capacity AND backlog are full — surface 429 with prior shape.
+        if execution_id:
+            db.update_execution_status(
+                execution_id=execution_id,
+                status=TaskExecutionStatus.FAILED,
+                error=(
+                    f"Agent at capacity ({max_parallel_tasks}/{max_parallel_tasks} parallel tasks running) "
+                    f"and backlog is full"
+                ),
+            )
+        idempotency_service.fail(idem)
+        raise ChatDispatchError(
+            429,
+            (
+                f"Agent '{name}' is at capacity ({max_parallel_tasks} parallel tasks) "
+                f"and its backlog is full. Try again later."
+            ),
+        )
+    except CircuitOpen as e:
+        # #526: dispatch breaker open — raised before the queue_persistent enqueue.
+        logger.warning(f"[{mode_label}] Agent '{name}' dispatch circuit open, rejecting")
+        idempotency_service.fail(idem)
+        raise _circuit_open_dispatch_error(name, execution_id, e)
+    except EphemeralBudgetExhausted as e:
+        # trinity-enterprise#69: ghost budget spent — 410, no enqueue.
+        idempotency_service.fail(idem)
+        raise _ephemeral_dispatch_error(name, execution_id, e)
+
+    return cap_result, effective_timeout
+
+
+def _map_task_failure(name, result):
+    """Shared /task failure translation (#679): a non-success terminal maps to
+    429 (at-capacity) / 504 (timed out) / 503. Raises ChatDispatchError."""
+    if result.status in ("failed", "cancelled"):
+        if "at capacity" in (result.error or ""):
+            raise ChatDispatchError(429, f"Agent '{name}' is at capacity. Try again later.")
+        elif "timed out" in (result.error or ""):
+            raise ChatDispatchError(504, result.error)
+        else:
+            raise ChatDispatchError(
+                503, result.error or "Failed to execute task. The agent may be unavailable."
+            )
+
+
+async def _dispatch_async(
+    *, request, name, current_user, execution_id, subscription_id,
+    collaboration_activity_id, self_task_activity_id, is_self_task, triggered_by,
+    reserved_event_dispatch, image_data, idem, x_source_agent, x_mcp_key_id, x_mcp_key_name,
+):
+    """Async branch (#95): pre-acquire capacity, then either report queued-202 or
+    spawn the background task and report accepted-202."""
+    cap_result, _effective_timeout = await _acquire_task_capacity(
+        mode_label="Task Async", name=name, request=request, execution_id=execution_id,
+        triggered_by=triggered_by, collaboration_activity_id=collaboration_activity_id,
+        is_self_task=is_self_task, self_task_activity_id=self_task_activity_id,
+        user_id=current_user.id, user_email=current_user.email or current_user.username,
+        subscription_id=subscription_id, x_source_agent=x_source_agent,
+        x_mcp_key_id=x_mcp_key_id, x_mcp_key_name=x_mcp_key_name, idem=idem,
+    )
+
+    if cap_result.state == "queued_persistent":
+        logger.info(
+            f"[Task Async] Agent '{name}' at capacity — execution {execution_id} queued to backlog"
+        )
+        _queued_payload = {
+            "status": "queued",
+            "execution_id": execution_id,
+            "agent_name": name,
+            "message": (
+                f"Agent at capacity; task queued. Poll GET "
+                f"/api/agents/{name}/executions/{execution_id} for results."
+            ),
+            "async_mode": True,
+        }
+        idempotency_service.complete(idem, execution_id, _queued_payload)
+        return _queued_payload
+
+    # Issue #279: done callback surfaces unhandled BG task exceptions.
+    def _on_task_done(task: asyncio.Task):
+        if task.cancelled():
+            logger.warning(f"[Task Async] Background task cancelled for agent '{name}', execution_id={execution_id}")
+        elif exc := task.exception():
+            logger.error(f"[Task Async] Unhandled exception in background task for agent '{name}', execution_id={execution_id}: {exc}")
+
+    bg_task = asyncio.create_task(
+        run_async_task(
+            agent_name=name,
+            request=request,
+            execution_id=execution_id,
+            collaboration_activity_id=collaboration_activity_id,
+            x_source_agent=x_source_agent,
+            user_id=current_user.id,
+            user_email=current_user.email or current_user.username,
+            subscription_id=subscription_id,
+            is_self_task=is_self_task,
+            self_task_activity_id=self_task_activity_id,
+            images=image_data,
+            dispatch_gate_checked=True,  # #526: router already gated at acquire()
+            triggered_by_override=(
+                RESERVED_EVENT_TRIGGER if reserved_event_dispatch else None
+            ),  # #1578 recursion-break
+        )
+    )
+    bg_task.add_done_callback(_on_task_done)
+
+    logger.info(f"[Task Async] Started background task for agent '{name}', execution_id={execution_id}")
+    _accepted_payload = {
+        "status": "accepted",
+        "execution_id": execution_id,
+        "agent_name": name,
+        "message": "Task accepted. Poll GET /api/agents/{name}/executions/{execution_id} for results.",
+        "async_mode": True,
+    }
+    idempotency_service.complete(idem, execution_id, _accepted_payload)
+    return _accepted_payload
+
+
+async def _dispatch_sync_backlog(*, name, execution_id, sync_effective_timeout, idem):
+    """Sync backlog long-poll (#498): wait for the drain terminal, or reconstruct
+    a minimal result from the row (non-drain terminal flip), translate failure,
+    and build the response. Side effects were handled inside the drain — not repeated."""
+    sync_wait_cap = 2 * sync_effective_timeout
+    logger.info(
+        f"[Task Sync] Agent '{name}' at capacity — execution {execution_id} "
+        f"queued to backlog; long-polling up to {sync_wait_cap}s"
+    )
+    try:
+        wait_payload = await wait_for_sync_terminal(execution_id, timeout=sync_wait_cap)
+    except asyncio.TimeoutError:
+        raise ChatDispatchError(
+            504,
+            (
+                f"Sync task on agent '{name}' did not complete within "
+                f"{sync_wait_cap}s. Execution {execution_id} may still be "
+                f"running; poll GET /api/agents/{name}/executions/{execution_id}."
+            ),
+        )
+
+    if wait_payload is not None and wait_payload.get("result") is not None:
+        result = wait_payload["result"]
+        sync_chat_session_id = wait_payload.get("chat_session_id")
+    else:
+        row = db.get_execution(execution_id)
+        if row is None:
+            raise ChatDispatchError(503, f"Execution {execution_id} disappeared while waiting")
+        from services.task_execution_service import TaskExecutionResult
+        result = TaskExecutionResult(
+            execution_id=execution_id,
+            status=row.status,
+            response=row.response or "",
+            cost=row.cost,
+            context_used=row.context_used,
+            context_max=row.context_max,
+            session_id=row.claude_session_id,
+            error=row.error,
+            raw_response={
+                "response": row.response or "",
+                "cost": row.cost,
+                "execution_id": execution_id,
+                "claude_session_id": row.claude_session_id,
+            },
+        )
+        sync_chat_session_id = None
+
+    _map_task_failure(name, result)
+
+    sync_response_data = result.raw_response or {}
+    if sync_chat_session_id:
+        sync_response_data["chat_session_id"] = sync_chat_session_id
+    sync_response_data["task_execution_id"] = execution_id
+    idempotency_service.complete(idem, execution_id, sync_response_data)
+    return sync_response_data
+
+
+async def _dispatch_sync_immediate(
+    *, request, name, current_user, execution_id, subscription_id, triggered_by,
+    collaboration_activity_id, image_data, idem, x_source_agent, x_mcp_key_id, x_mcp_key_name,
+):
+    """Sync immediate path (EXEC-024): delegate to the single applier
+    (task_execution_service.execute_task), complete the collaboration activity,
+    translate failure, persist to the chat session (#1444), and build the response."""
+    task_execution_service = get_task_execution_service()
+    result = await task_execution_service.execute_task(
+        agent_name=name,
+        message=request.message,
+        triggered_by=triggered_by,
+        source_user_id=current_user.id,
+        source_user_email=current_user.email or current_user.username,
+        source_agent_name=x_source_agent,
+        source_mcp_key_id=x_mcp_key_id,
+        source_mcp_key_name=x_mcp_key_name,
+        model=request.model,
+        timeout_seconds=request.timeout_seconds,  # TIMEOUT-001: None = use agent's config
+        resume_session_id=request.resume_session_id,
+        allowed_tools=request.allowed_tools,
+        system_prompt=request.system_prompt,
+        execution_id=execution_id,
+        slot_already_held=True,  # Issue #498: router pre-acquired
+        images=image_data,
+        dispatch_gate_checked=True,  # #526: router already gated at acquire()
+    )
+
+    if collaboration_activity_id:
+        await activity_service.complete_activity(
+            activity_id=collaboration_activity_id,
+            # #1332: a cancelled collaboration turn reads as CANCELLED, not FAILED.
+            status=activity_state_for_terminal(result.status),
+            details={
+                "response_length": len(result.response),
+                "execution_id": execution_id,
+            },
+            error=result.error if result.status == TaskExecutionStatus.FAILED else None,
+        )
+
+    _map_task_failure(name, result)
+
+    response_data = result.raw_response
+
+    # Persist to chat session if requested (#1444: guarded on SUCCESS).
+    if request.save_to_session:
+        chat_session_id = await chat_persistence_service.persist_chat_session(
+            agent_name=name,
+            request=request,
+            result=result,
+            user_id=current_user.id,
+            user_email=current_user.email or current_user.username,
+            subscription_id=subscription_id,
+        )
+        if chat_session_id:
+            response_data["chat_session_id"] = chat_session_id
+        elif result.status == TaskExecutionStatus.SUCCESS:
+            # Fail-loud (#1444): never 500 a completed, billed turn.
+            response_data["chat_persist_failed"] = True
+
+    response_data["task_execution_id"] = execution_id
+    idempotency_service.complete(idem, execution_id, response_data)
+    return response_data
+
+
+async def _dispatch_sync(
+    *, request, name, current_user, execution_id, subscription_id,
+    collaboration_activity_id, self_task_activity_id, is_self_task, triggered_by,
+    image_data, idem, x_source_agent, x_mcp_key_id, x_mcp_key_name,
+):
+    """Sync branch (#498): pre-acquire; on immediate admit run the single
+    applier, else long-poll the backlog drain."""
+    cap_result, sync_effective_timeout = await _acquire_task_capacity(
+        mode_label="Task Sync", name=name, request=request, execution_id=execution_id,
+        triggered_by=triggered_by, collaboration_activity_id=collaboration_activity_id,
+        is_self_task=is_self_task, self_task_activity_id=self_task_activity_id,
+        user_id=current_user.id, user_email=current_user.email or current_user.username,
+        subscription_id=subscription_id, x_source_agent=x_source_agent,
+        x_mcp_key_id=x_mcp_key_id, x_mcp_key_name=x_mcp_key_name, idem=idem,
+    )
+
+    if cap_result.state != "admitted":
+        return await _dispatch_sync_backlog(
+            name=name, execution_id=execution_id,
+            sync_effective_timeout=sync_effective_timeout, idem=idem,
+        )
+    return await _dispatch_sync_immediate(
+        request=request, name=name, current_user=current_user, execution_id=execution_id,
+        subscription_id=subscription_id, triggered_by=triggered_by,
+        collaboration_activity_id=collaboration_activity_id, image_data=image_data,
+        idem=idem, x_source_agent=x_source_agent, x_mcp_key_id=x_mcp_key_id,
+        x_mcp_key_name=x_mcp_key_name,
+    )
+
+
+async def dispatch_parallel_task(
+    *, request, name, current_user, container, x_source_agent, x_via_mcp,
+    x_mcp_key_id, x_mcp_key_name, idempotency_key, x_event_trigger, x_internal_secret,
+):
+    """The /task dispatch orchestrator (Invariant #1). Owns derive → idempotency
+    (via dispatch_admission_service) → file upload → create-row+activities →
+    async/sync fork. Returns a response dict OR a ChatAdmissionReplay (router maps
+    to 200/409); raises ChatDispatchError for the HTTP error cases. The router
+    keeps the resume-validation + #1068 timeout guards ahead of this call (their
+    order + the redis-via-router #1068 helper are boundary concerns — RD10)."""
+    derivation = derive_source_and_trigger(
+        name=name, x_source_agent=x_source_agent, x_via_mcp=x_via_mcp,
+        x_event_trigger=x_event_trigger, x_internal_secret=x_internal_secret,
+        current_user=current_user,
+    )
+
+    # RELIABILITY-006 (#525): idempotency begin/replay (shared with /chat, RD2).
+    idem, replay = dispatch_admission_service.begin_task_idempotency(
+        name=name, idempotency_key=idempotency_key,
+    )
+    if replay is not None:
+        await dispatch_admission_service.audit_idempotent_replay(
+            name=name, endpoint=f"/api/agents/{name}/task", x_via_mcp=x_via_mcp,
+            x_source_agent=x_source_agent, x_mcp_key_id=x_mcp_key_id,
+            x_mcp_key_name=x_mcp_key_name, current_user=current_user,
+            idempotency_key=idempotency_key, idem=idem,
+        )
+        return replay
+
+    image_data = await process_task_file_uploads(
+        request=request, name=name, container=container, current_user=current_user,
+    )
+
+    (
+        execution_id, subscription_id, collaboration_activity_id, self_task_activity_id,
+    ) = await create_task_execution_and_activities(
+        request=request, name=name, current_user=current_user,
+        x_source_agent=x_source_agent, x_mcp_key_id=x_mcp_key_id,
+        x_mcp_key_name=x_mcp_key_name, triggered_by=derivation.triggered_by,
+        is_self_task=derivation.is_self_task, idem=idem,
+    )
+
+    if request.async_mode:
+        return await _dispatch_async(
+            request=request, name=name, current_user=current_user, execution_id=execution_id,
+            subscription_id=subscription_id, collaboration_activity_id=collaboration_activity_id,
+            self_task_activity_id=self_task_activity_id, is_self_task=derivation.is_self_task,
+            triggered_by=derivation.triggered_by, reserved_event_dispatch=derivation.reserved_event_dispatch,
+            image_data=image_data, idem=idem, x_source_agent=x_source_agent,
+            x_mcp_key_id=x_mcp_key_id, x_mcp_key_name=x_mcp_key_name,
+        )
+    return await _dispatch_sync(
+        request=request, name=name, current_user=current_user, execution_id=execution_id,
+        subscription_id=subscription_id, collaboration_activity_id=collaboration_activity_id,
+        self_task_activity_id=self_task_activity_id, is_self_task=derivation.is_self_task,
+        triggered_by=derivation.triggered_by, image_data=image_data, idem=idem,
+        x_source_agent=x_source_agent, x_mcp_key_id=x_mcp_key_id, x_mcp_key_name=x_mcp_key_name,
+    )
