+++ tests/unit/test_1578_task_completion_events.py
@@ -705,6 +705,7 @@ class TestRouterRecursionBreakGate:
         from fastapi import HTTPException
         from routers.chat import execute_parallel_task
         import services.dispatch_admission_service as dispatch
+        import services.chat_execution_service as ce
         from models import ParallelTaskRequest
         from services.capacity_manager import CapacityFull
 
@@ -737,14 +738,15 @@ class TestRouterRecursionBreakGate:
         with (
             patch.dict(os.environ, {"INTERNAL_API_SECRET": self._SECRET}),
             patch.object(chat, "get_agent_container", return_value=container),
-            patch.object(chat, "idempotency_service", isvc),
+            patch.object(ce, "idempotency_service", isvc),
             patch.object(dispatch, "idempotency_service", isvc),
-            patch.object(chat, "dispatch_breaker_active", return_value=False),
-            patch.object(chat, "get_capacity_manager", return_value=cap),
+            patch.object(ce, "dispatch_breaker_active", return_value=False),
+            patch.object(ce, "get_capacity_manager", return_value=cap),
             patch.object(
-                chat, "activity_service",
+                ce, "activity_service",
                 MagicMock(track_activity=AsyncMock(return_value="act1")),
             ),
+            patch.object(ce, "db", mock_db),
             patch.object(chat, "db", mock_db),
         ):
             with pytest.raises(HTTPException):  # CapacityFull → 429 after create
